Move a file, fix errors after merging and moving
